+++ examples/resources/nxos_ipv4_access_list_entry/resource.tf
@@ -5,10 +5,10 @@ resource "nxos_ipv4_access_list_entry" "example" {
   action                    = "permit"
   dscp                      = 0
   destination_address_group = "AG1"
-  destination_port_1        = 443
-  destination_port_2        = 0
+  destination_port_1        = "443"
+  destination_port_2        = "0"
   destination_port_group    = "PG1"
-  destination_port_mask     = 0
+  destination_port_mask     = "0"
   destination_port_operator = "eq"
   destination_prefix        = "10.1.1.0"
   destination_prefix_length = "24"
@@ -32,10 +32,10 @@ resource "nxos_ipv4_access_list_entry" "example" {
   rev                       = false
   rst                       = false
   source_address_group      = "AG2"
-  source_port_1             = 443
-  source_port_2             = 0
+  source_port_1             = "443"
+  source_port_2             = "0"
   source_port_group         = "PG2"
-  source_port_mask          = 0
+  source_port_mask          = "0"
   source_port_operator      = "eq"
   source_prefix             = "20.1.0.0"
   source_prefix_length      = "16"