+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Add nxos_ipv4_access_list resource and data source
 - Add nxos_ipv4_access_list_entry resource and data source
+- Add nxos_ipv4_access_list_policy_ingress_interface resource and data source
+- Add nxos_ipv4_access_list_policy_ingress_interface_instance resource and data source
 
 ## 0.3.19